tech(Alert): add AlertBase component with implementation of alert modal
