+++ packages/vkui/src/components/Alert/Alert.tsx
@@ -1,13 +1,7 @@
 'use client';
 
 import * as React from 'react';
-import { Icon20Cancel } from '@vkontakte/icons';
-import { classNames, hasReactNode, noop } from '@vkontakte/vkjs';
-import { useAdaptivityWithJSMediaQueries } from '../../hooks/useAdaptivityWithJSMediaQueries';
 import { type UseFocusTrapProps } from '../../hooks/useFocusTrap';
-import { usePlatform } from '../../hooks/usePlatform';
-import { useCSSKeyframesAnimationController } from '../../lib/animation';
-import { stopPropagation } from '../../lib/utils';
 import { warnOnce } from '../../lib/warnOnce';
 import type {
   AlignType,
@@ -19,14 +13,9 @@ import type { AppRootPortalProps } from '../AppRoot/AppRootPortal';
 import { AppRootPortal } from '../AppRoot/AppRootPortal';
 import { useScrollLock } from '../AppRoot/ScrollContext';
 import type { ButtonProps } from '../Button/Button';
-import { FocusTrap } from '../FocusTrap/FocusTrap';
-import { IconButton } from '../IconButton/IconButton';
-import { ModalDismissButton } from '../ModalDismissButton/ModalDismissButton';
 import { PopoutWrapper } from '../PopoutWrapper/PopoutWrapper';
 import type { AlertActionProps } from './AlertAction';
-import { AlertActions } from './AlertActions';
-import { AlertDescription, AlertTitle } from './AlertTypography';
-import styles from './Alert.module.css';
+import { AlertBase } from './AlertBase';
 
 type AlertActionMode = 'cancel' | 'destructive' | 'default';
 
@@ -96,75 +85,23 @@ const warn = warnOnce('Alert');
  * @see https://vkcom.github.io/VKUI/#/Alert
  */
 export const Alert = ({
-  actions,
-  actionsLayout = 'horizontal',
-  children,
-  className,
-  style,
-  title,
-  description,
-  onClose,
-  dismissLabel = 'Закрыть предупреждение',
-  renderAction,
-  actionsAlign,
-  dismissButtonMode = 'outside',
-  dismissButtonTestId,
-  getRootRef,
   usePortal,
-  onClick,
-  allowClickPropagation = false,
-  titleTestId,
-  descriptionTestId,
+  style,
+  className,
+  getRootRef,
   ...restProps
 }: AlertProps): React.ReactNode => {
-  const generatedId = React.useId();
-
-  const titleId = `vkui-alert-${generatedId}-title`;
-  const descriptionId = `vkui-alert-${generatedId}-description`;
-
-  const platform = usePlatform();
-  const { isDesktop } = useAdaptivityWithJSMediaQueries();
-
   const [closing, setClosing] = React.useState(false);
-  const itemActionCallbackRef = React.useRef(noop);
-  const [animationState, animationHandlers] = useCSSKeyframesAnimationController(
-    closing ? 'exit' : 'enter',
-    {
-      onExited() {
-        itemActionCallbackRef.current();
-        itemActionCallbackRef.current = noop;
-        onClose();
-      },
-    },
-  );
-  const isDismissButtonVisible = isDesktop && platform !== 'ios';
-  const elementRef = React.useRef<HTMLDivElement>(null);
 
   const close = React.useCallback(() => {
     setClosing(true);
   }, []);
 
-  const onItemClick = React.useCallback(
-    (item: AlertActionInterface) => {
-      const { action: itemAction, autoCloseDisabled = false } = item;
-
-      if (autoCloseDisabled) {
-        itemAction && itemAction({ close });
-      } else {
-        if (itemAction) {
-          itemActionCallbackRef.current = itemAction;
-        }
-        setClosing(true);
-      }
-    },
-    [close],
-  );
-
   useScrollLock();
 
   if (
     process.env.NODE_ENV === 'development' &&
-    !title &&
+    !restProps.title &&
     !restProps['aria-label'] &&
     !restProps['aria-labelledby']
   ) {
@@ -172,13 +109,6 @@ export const Alert = ({
       'Если "title" не используется, то необходимо задать либо "aria-label", либо "aria-labelledby" (см. правило axe aria-dialog-name)',
     );
   }
-
-  const handleClick = allowClickPropagation
-    ? onClick
-    : (event: React.MouseEvent<HTMLElement>) => {
-        stopPropagation(event);
-        onClick?.(event);
-      };
 
   return (
     <AppRootPortal usePortal={usePortal}>
@@ -190,68 +120,7 @@ export const Alert = ({
         getRootRef={getRootRef}
         strategy="fixed"
       >
-        <FocusTrap
-          {...animationHandlers}
-          onClick={handleClick}
-          getRootRef={elementRef}
-          onClose={close}
-          autoFocus={animationState === 'entered'}
-          className={classNames(
-            styles.host,
-            platform === 'ios' && styles.ios,
-            platform === 'vkcom' && styles.vkcom,
-            closing ? styles.closing : styles.opening,
-            isDesktop && styles.desktop,
-          )}
-          role="alertdialog"
-          aria-modal
-          aria-labelledby={titleId}
-          aria-describedby={descriptionId}
-          {...restProps}
-        >
-          <div
-            className={classNames(
-              styles.content,
-              dismissButtonMode === 'inside' && styles.contentWithButton,
-            )}
-          >
-            {hasReactNode(title) && (
-              <AlertTitle data-testid={titleTestId} id={titleId}>
-                {title}
-              </AlertTitle>
-            )}
-            {hasReactNode(description) && (
-              <AlertDescription data-testid={descriptionTestId} id={descriptionId}>
-                {description}
-              </AlertDescription>
-            )}
-            {children}
-            {isDismissButtonVisible && dismissButtonMode === 'inside' && (
-              <IconButton
-                label={dismissLabel}
-                className={classNames(styles.dismiss, 'vkuiInternalAlert__dismiss')}
-                onClick={close}
-                hoverMode="opacity"
-                activeMode="opacity"
-                data-testid={dismissButtonTestId}
-              >
-                <Icon20Cancel />
-              </IconButton>
-            )}
-          </div>
-          {isDismissButtonVisible && dismissButtonMode === 'outside' && (
-            <ModalDismissButton onClick={close} data-testid={dismissButtonTestId}>
-              {dismissLabel}
-            </ModalDismissButton>
-          )}
-          <AlertActions
-            actions={actions}
-            actionsAlign={actionsAlign}
-            actionsLayout={actionsLayout}
-            renderAction={renderAction}
-            onItemClick={onItemClick}
-          />
-        </FocusTrap>
+        <AlertBase {...restProps} closing={closing} setClosing={setClosing} />
       </PopoutWrapper>
     </AppRootPortal>
   );
